1
2

+++ 4.0.1/aspnet-core/src/Application/Goods/Brand/BrandAppService.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using Abp.Application.Services;
 using Abp.Domain.Repositories;
 using Abp.Runtime.Session;
+using Microsoft.AspNetCore.Mvc;
 using UnionMall.IRepositorySql;
 
 namespace UnionMall.Goods.Brand
@@ -22,39 +23,48 @@ namespace UnionMall.Goods.Brand
             _Repository = Repository;
             _AbpSession = AbpSession;
         }
-        public async Task CreateOrEditAsync(Brand dto)
+        public async Task<JsonResult> CreateOrEditAsync(Brand dto)
         {
             if (dto.Id > 0)
             {
-                var query = _Repository.Get(dto.Id);
+                var query = _Repository.FirstOrDefaultAsync(c => c.Title == dto.Title);
                 if (query == null)
                 {
-
+                    return new JsonResult(new { succ = false, msg = "NotExist" });
                 }
                 await _Repository.UpdateAsync(dto);
+                return new JsonResult(new { succ = true, msg = "" });
             }
             else
             {
                 var query = _Repository.FirstOrDefault(c => c.Title == dto.Title);
                 if (query != null)
                 {
-
+                    return new JsonResult(new { succ = false, msg = "{0}AlreadyExist" });
                 }
 
                 if (_AbpSession.TenantId != null)
                 { dto.TenantId = (int)_AbpSession.TenantId; }
                 await _Repository.InsertAsync(dto);
+                return new JsonResult(new { succ = true, msg = "" });
             }
 
         }
 
-        public async Task DeleteAsync(long id)
+        public async Task<JsonResult> DeleteAsync(long id)
         {
+            var query = _Repository.FirstOrDefault(c => c.Id == id);
+            if (query == null)
+            {
+                return new JsonResult(new { succ = true, msg = "NotExist" });
+            }
             await _Repository.DeleteAsync(id);
+            return new JsonResult(new { succ = true, msg = "" });
         }
 
         public async Task<Brand> GetByIdAsync(long Id)
         {
+
             var s = await _Repository.GetAsync(Id);
             return s;
         }
